Compiling using "gcc version 4.6.1 (Ubuntu/Linaro 4.6.1-9ubuntu3)" is working again

+++ Plugins/PLFrontendQt/src/DockWidget/DockWidgetSceneGraphQObject.cpp
@@ -83,7 +83,7 @@ void DockWidgetSceneGraphQObject::QtSlotTreeViewDoubleClicked(const QModelIndex 
 
 void DockWidgetSceneGraphQObject::QtSlotFilterTypeChanged(int filterId)
 {
-	m_pDockWidgetSceneGraph->m_pSortAndFilterModel->setFilterRole((filterId == ByClassName) ? DataModels::SceneGraphTreeModel::ClassNameRole : Qt::DisplayRole);
+	m_pDockWidgetSceneGraph->m_pSortAndFilterModel->setFilterRole((filterId == ByClassName) ? DataModels::SceneGraphTreeModel::ClassNameRole : static_cast<int>(Qt::DisplayRole));
 }
 
 void DockWidgetSceneGraphQObject::QtSlotFilterChanged(const QString &text)
